Render all wires on a single SVG layer
Uses multiple <path>s to preserve React's DOM optimizations

Currently the <svg> is statically 10k pixels wide and tall. This number is arbitrary, but it turns out that without infinite-scrolling implemented this makes it much easier to scroll around the grid and move nodes off-screen. Still, this shouldn't be shipped if a better solution can be accomplished reasonably.

Closes #4