Another mass restructuring
This came about from a stronger understanding of what a "node" is in graph theory, and how to model the user-facing
graph-like structure using a _real_ graph. The biggest underlying change is that nodes are now a single "point" in the
UI, represented by a pin on the side of a card or graph boundary. This simplifies algorithms and iteration-based
rendering.

- Top-level data types removed in favor of more structured hierarchy
- "Nodes" become "cards", "node" now refers only to nodes in the underlying graph
- Cards are rendered to the graph view with pins for each node
- "NodePrototype"s are now "CardTemplate"s, "NodeInfo" is now "CardSnapshot"
- "Palette" contains both the maps of card templates and their directory structure in the document
- Layout information is stored in the items being layed out, in particular cells and cards store their own locations

Runtime errors still prevent rendering at this commit.