Further restructuring and normalizing
Decided to go all the way and include database-style joint "tables" and remove reference duplication. Finding, say, all
the nodes for a property requires only a single loop, and can be combined with other filters if they exist, so
performance is not an immediate concern.

This allows the reduction of the various `NodeSource` descendent types down to just a `Source` type describing the
inputs and outputs, since child data (nodes, rows, etc) can be related to them through other tables. `Source` and
`Directory` both satisfy the `NamedEntity` interface, so they can be handled identically by tree search and rendering.

`Pin`s are now "Terminals", although no single interface exists for them. Instead, things which point to a Terminal use a
`TerminalReference`, and Terminal metadata is stored in `TerminalDescription`s.

`Cell`s and `Sheet`s have gotten minor tweaks to sidestep compiler errors, but are still very neglected. They'll
probably need to be rewritten when it comes time to render them.